requirements.txt
Django==5.2.6
Pillow==11.3.0
qrcode[pil]==8.2
django-cors-headers==4.7.0